add qwt package
Encapsulate source file modifications into a SubPatch object which can be applied to a Package.  Add a dictionary of variable settings which are substituted into all patches and commands using % operator.  Override QMAKE_MOC on the qwt qmake command line so it is not necessary to edit the MOC path in the Makefiles generated by qmake.